+++ src/components/DashBoard.tsx
@@ -34,7 +34,7 @@ import ProposalCreate from './proposal/ProposalCreate';
 import ProposalEdit from './proposal/ProposalEdit';
 import ProposalPage from './proposal/ProposalPage';
 import InstrSciUpcomingExperimentTimesTable from './proposalBooking/InstrSciUpcomingExperimentTimesTable';
-import UserExperimentTimesTable from './proposalBooking/UserExperimentTimesTable';
+import UserExperimentTimesTable from './proposalBooking/UserExperimentsTable';
 import ProposalTableReviewer from './review/ProposalTableReviewer';
 import SampleSafetyPage from './sample/SampleSafetyPage';
 import SEPPage from './SEP/SEPPage';
